requirements.txt
fastapi==0.104.1  
uvicorn==0.23.2  
langchain==0.1.14  
langchain-community==0.0.14  
langchain-huggingface==0.1.2  
psycopg2-binary==2.9.9  
transformers==4.37.2  
sentence-transformers==2.2.2  
torch==2.1.2  
pydantic==2.4.2  
python-dotenv==1.0.1  
